# public/openapi/chatgpt-actions-v1.yaml
...
openapi: 3.1.0
info:
  title: Mission Control ChatGPT Actions API
  version: 1.0.0
  description: 'Action surface for the private Mission Control custom GPT. This schema
    exposes the non-destructive v1 read/write operations needed for briefs, notes
    parsing, task management, focus, projects, sprints, stakeholders, and commitments.

'
servers:
- url: https://mission-control-orpin-chi.vercel.app
security:
- MissionControlKey: []
...
paths:
  "/api/calendar":
    get:
      tags:
      - Calendar
      operationId: get_calendar
      summary: Get calendar events for a date range
      parameters:
      - name: rangeStart
        in: query
        required: true
        schema:
          type: string
          format: date
        description: Inclusive start date in YYYY-MM-DD.
      - name: rangeEnd
        in: query
        required: true
        schema:
          type: string
          format: date
        description: Inclusive end date in YYYY-MM-DD.
      responses:
        '200':
          description: Calendar snapshot
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CalendarResponse"
        default:
          "$ref": "#/components/responses/ErrorResponse"
...
  "/api/planner/sync-today":
    post:
      tags:
      - Planner
      operationId: sync_today
      summary: Promote the approved today task list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - task_ids
              properties:
                task_ids:
                  type: array
                  minItems: 1
                  maxItems: 20
                  items:
                    type: string
                    format: uuid
              additionalProperties: false
      responses:
        '200':
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  promoted:
                    type: integer
                  demoted:
                    type: integer
                  skipped_pinned:
                    type: integer
                  sync_at:
                    type: string
                    format: date-time
                additionalProperties: false
        default:
          "$ref": "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    MissionControlKey:
      type: apiKey
      in: header
      name: X-Mission-Control-Key
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/BasicError"
  schemas:
    BasicError:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
      required:
      - error
      additionalProperties: true
...
    CalendarResponse:
      type: object
      properties:
        rangeStart:
          type: string
          format: date
        rangeEnd:
          type: string
          format: date
        source:
          type: string
        warning:
          anyOf:
          - type: string
          - type: 'null'
        warnings:
          type: array
          items:
            type: string
        ingest:
          type: object
          properties:
            source:
              type: string
            ingestedCount:
              type: integer
            warningCount:
              type: integer
          additionalProperties: true
        events:
          type: array
          items:
            type: object
            properties:
              start_at:
                type: string
                format: date-time
              end_at:
                type: string
                format: date-time
              title:
                type: string
              with_display:
                type: array
                items:
                  type: string
              body_scrubbed_preview:
                anyOf:
                - type: string
                - type: 'null'
              is_all_day:
                type: boolean
              external_event_id:
                type: string
              meeting_context:
                anyOf:
                - type: string
                - type: 'null'
              date_et:
                type: string
              start_time_et:
                type: string
              end_time_et:
                type: string
              time_range_et:
                type: string
              temporal_status:
                type: string
                enum:
                - past
                - in_progress
                - upcoming
            additionalProperties: true
        busyBlocks:
          type: array
          items:
            type: object
            additionalProperties: true
        stats:
          type: object
          additionalProperties: true
        changesSince:
          type: object
          additionalProperties: true
        generatedAt:
          type: string
          format: date-time
      required:
      - rangeStart
      - rangeEnd
      - events
      additionalProperties: true
...